+++ CleanArchExample/data/src/main/kotlin/com/example/data/datasource/dayofweek/DayOfWeekDataStore.kt
@@ -3,28 +3,10 @@ package com.example.data.datasource.dayofweek
 import com.example.domain.dayofweek.entity.DayOfWeek
 import com.example.domain.dayofweek.entity.Language
 import com.example.domain.repository.DayOfWeekRepository
-import java.util.*
 import javax.inject.Inject
 
 class DayOfWeekDataStore @Inject constructor() : DayOfWeekRepository {
-    override fun getTodaysDayOfWeek(): DayOfWeek {
-        val calendar = Calendar.getInstance(TimeZone.getTimeZone("Asia/Tokyo"), Locale.JAPAN)
-//        val calendar = Calendar.getInstance(TimeZone.getTimeZone("Europe/Berlin"), Locale.GERMANY)
-
-        calendar.time = Date()
-        return when (calendar.get(Calendar.DAY_OF_WEEK)) {
-            Calendar.SUNDAY -> DayOfWeek.SUNDAY
-            Calendar.MONDAY -> DayOfWeek.MONDAY
-            Calendar.TUESDAY -> DayOfWeek.TUESDAY
-            Calendar.WEDNESDAY -> DayOfWeek.WEDNESDAY
-            Calendar.THURSDAY -> DayOfWeek.THURSDAY
-            Calendar.FRIDAY -> DayOfWeek.FRIDAY
-            Calendar.SATURDAY -> DayOfWeek.SATURDAY
-            else -> throw IllegalArgumentException()
-        }
-    }
-
-    override fun getDayOfWeekWords(language: Language): Map<DayOfWeek, String> {
+    override fun getDayOfWeekMap(language: Language): Map<DayOfWeek, String> {
         return when (language) {
             Language.English ->
                 mapOf(
