+++ CleanArchExample/data/src/main/kotlin/com/example/data/utils/di/RepositoryModules.kt
@@ -1,7 +1,9 @@
 package com.example.data.utils.di
 
 import com.example.data.datasource.dayofweek.DayOfWeekDataStore
+import com.example.data.datasource.system.SystemDataStore
 import com.example.domain.repository.DayOfWeekRepository
+import com.example.domain.repository.SystemRepository
 import dagger.Binds
 import dagger.Module
 
@@ -12,4 +14,7 @@ abstract class RepositoryModules {
 //  @Binds abstract fun bindSomeRepository(dataStore: SomeDataStore): SomeRepository
     @Binds
     abstract fun bindDayOfWeekRepository(dataStore: DayOfWeekDataStore): DayOfWeekRepository
+
+    @Binds
+    abstract fun bindSystemRepository(dataStore: SystemDataStore): SystemRepository
 }
