+++ CleanArchExample/domain/src/main/kotlin/com/example/domain/repository/SystemRepository.kt
@@ -1,0 +1,7 @@
+package com.example.domain.repository
+
+import com.example.domain.dayofweek.entity.DayOfWeek
+
+interface SystemRepository {
+    fun getTodaysDayOfWeek(): DayOfWeek
+}
